+++ Runtime/Zenject/ILocalizableObjectFactory.cs
@@ -1,9 +1,7 @@
-using UnityEngine;
-
 namespace Playdarium.Localization.Runtime.Zenject
 {
 	public interface ILocalizableObjectFactory
 	{
-		ILocalizableObject Create(LocalizableField field, Component localizable);
+		ILocalizableObject Create(object localizable, ILocalizationSettings settings);
 	}
 }