+++ Runtime/Zenject/ILocalizationInjector.cs
@@ -5,6 +5,8 @@ namespace Playdarium.Localization.Runtime.Zenject
 {
 	public interface ILocalizationInjector
 	{
-		void Localize(Component localizable, List<ILocalizableObject> localizableObjects);
+		void Localize(Component localizableView, List<ILocalizableObject> localizableObjects);
+
+		ILocalizableObject Localize(object localizable, ILocalizationSettings settings);
 	}
 }